backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
llama-cpp-python[server]
supabase-py
pydantic-settings
stripe
psycopg2-binary
